之前的石门相关名字RockGate修改为钻场相关的名字DrillSite 但是反应器相关的名字未修改
