+++ ARX_RCU/ARX_RCU.vcproj
@@ -457,10 +457,6 @@
 					RelativePath=".\RcuEditOpenPoreDlg.cpp"
 					>
 				</File>
-				<File
-					RelativePath=".\RcuEditRockGateDlg.cpp"
-					>
-				</File>
 			</Filter>
 			<Filter
 				Name="reactor"
@@ -544,10 +540,6 @@
 				</File>
 				<File
 					RelativePath=".\RcuEditOpenPoreDlg.h"
-					>
-				</File>
-				<File
-					RelativePath=".\RcuEditRockGateDlg.h"
 					>
 				</File>
 			</Filter>
