+++ ARX_RCU/RcuDataLink.h
@@ -3,45 +3,6 @@
 #include "../MineGE/DataLink.h"
 
 //4类石门揭煤图元数据关联,减少不必要的字符串与其他数据类型的转换代码
-
-//class RockGateLink : public DataLink
-//{
-//public:
-//	RockGateLink() : m_length(0), m_width(0), m_height(0), 
-//		             m_left(0), m_right(0), m_top(0), m_bottom(0),
-//					 m_dist(0)
-//	{
-//
-//	}
-//protected:
-//	virtual void regDatas()
-//	{
-//		linkStringData( _T( "名称" ), &m_name );
-//		linkDoubleData( _T( "长度" ), &m_length );
-//		linkDoubleData( _T( "宽度" ), &m_width );
-//		linkDoubleData( _T( "高度" ), &m_height );
-//		linkDoubleData( _T( "最小法距" ), &m_dist );
-//		linkPointData( _T( "迎头底板中心点坐标" ), &m_pt );
-//		linkDoubleData( _T( "左帮保护" ), &m_left );
-//		linkDoubleData( _T( "右帮保护" ), &m_right );
-//		linkDoubleData( _T( "上帮保护" ), &m_top );
-//		linkDoubleData( _T( "下帮保护" ), &m_bottom );
-//	}
-//
-//public:
-//	//坐标内部存储是用字符串表示的,xyz之间通过逗号分隔
-//	AcGePoint3d m_pt;
-//	double m_length;
-//	double m_width;
-//	double m_height;
-//	double m_left;
-//	double m_right;
-//	double m_top;
-//	double m_bottom;
-//	double m_dist;
-//	CString m_name;
-//};
-
 class CoalSurfaceLink : public DataLink
 {
 public:
