+++ ARX_RCU/RcuHelper.h
@@ -6,7 +6,7 @@ class RcuHelper
 {
 public:
 	//查找所有的钻场
-	static void FindAllRockGates(AcDbObjectIdArray& rock_gates);
+	static void FindAllDrillSite(AcDbObjectIdArray& rock_gates);
 	//查找与钻场关联的煤层
 	static void GetRelatedCoalSurface(const AcDbObjectId& drill_site, AcDbObjectId& coal_surf);
 	////查找与钻场关联的钻场
@@ -28,7 +28,7 @@ public:
 	static bool ReadPoreData(const AcDbObjectId& pore, PoreLink& po_link);
 
 	//得到钻场以及关联煤层的数据
-	static bool GetRockGateAndCoalSurfDatas(const AcDbObjectId& drill_site, DrillSiteLink& ds_link, CoalSurfaceLink& cs_link);
+	static bool GetDrillSiteAndCoalSurfDatas(const AcDbObjectId& drill_site, DrillSiteLink& ds_link, CoalSurfaceLink& cs_link);
 	//得到钻场的数据
 	static bool GetDrillSiteData(const AcDbObjectId& drill_site, DrillSiteLink& ds_link);
 
@@ -54,14 +54,14 @@ public:
 	//根据钻场数据修改关联的图元(开孔)
 	static bool ModifyDrillSiteRelatedGEs(const AcDbObjectId& drill_site, DrillSiteLink& ds_link);
 	//根据钻场数据修改关联的图元(煤层、钻场)
-	static bool ModifyRockGateRelatedGEs(const AcDbObjectId& drill_site, DrillSiteLink& ds_link, CoalSurfaceLink& cs_link);
+	static bool ModifyDrillSiteRelatedGEs(const AcDbObjectId& drill_site, DrillSiteLink& ds_link, CoalSurfaceLink& cs_link);
 	//根据煤层数据修改关联的图元(终孔)
 	static bool ModifyCoalSurfRelatedGEs(const AcDbObjectId& coal_surf, CoalSurfaceLink& cs_link);
 
 	//修改钻场参数
 	static bool ModifyDrillSitePt(const AcDbObjectId& drill_site, DrillSiteLink& ds_link);
 	//获取钻场的插入点坐标
-	static bool GetRockGateInsertPt(const AcDbObjectId& drill_site, AcGePoint3d& insertPt);
+	static bool GetDrillSiteInsertPt(const AcDbObjectId& drill_site, AcGePoint3d& insertPt);
 	//获取煤层的插入点坐标
 	static bool GetCoalSurfInsertPt(const AcDbObjectId& coal_surf, AcGePoint3d& insertPt);
 	//设置钻场图元的2点坐标
